fix(dataset): store full label mask as int8 and cast labels at loss time
Agent-Logs-Url: https://github.com/TEAT-Unicaen/R.A.I.D.-RAM-Artificial-Inspection-Dump/sessions/186c6d7c-96f0-46f3-bcfc-3c4a799aae97

+++ src/dumpManager/RamDumpDataset.py
@@ -58,7 +58,7 @@ class RamDumpDataset(Dataset):
         print(f"Building full label mask ({bin_size / 1e6:.1f}MB)...", flush=True)
         
         # Initialize with -1 (padding/unlabeled)
-        self.full_label_mask = np.full(bin_size, -1, dtype=np.float32)
+        self.full_label_mask = np.full(bin_size, -1, dtype=np.int8)
         
         # Fill in labels from metadata in single pass
         for entry in self.metadata:
